fix: audit batch 1 — 20 bug fixes, security hardening, and UI improvements (#355)
* fix: use float division for engine volume display (#341)

Integer division l.EngineVolume/1000 truncates small engines
(e.g. 999cc shows as 0L). Use float64 division instead.

Closes #341

Made-with: Cursor

* fix: add Vary: Origin header to CORS middleware (#316)

Without Vary: Origin, shared caches may serve a CORS response
intended for one origin to a different origin, causing cache
poisoning.

Closes #316

Made-with: Cursor

* fix: ProxyPool.Next fallback resets proxy with oldest cooldown (#320)

The fallback used p.idx which had already been incremented past the
last checked proxy. Now picks the proxy whose cooldown expired
earliest, which is the safest choice when all proxies are unhealthy.

Closes #320

Made-with: Cursor

* fix: CircuitBreaker ignores context cancellation errors (#321)

context.Canceled and context.DeadlineExceeded during graceful
shutdown are not real failures and should not trip the breaker.

Closes #321

Made-with: Cursor

* fix: replace time.After with reusable timer in scheduler loop (#319)

time.After creates a new timer each iteration that cannot be
garbage collected until it fires. Under high trigger frequency
this leaks memory. Reuse a single timer with proper Stop/Reset.

Closes #319

Made-with: Cursor

* fix: applyDefaults uses new slice instead of [:0] append trick (#330)

The [:0] append trick reuses the backing array, which can cause
subtle aliasing bugs if the original slice is referenced elsewhere.
Allocate a fresh slice instead.

Closes #330

Made-with: Cursor

* fix: validate CORS origins have scheme and host (#336)

url.Parse accepts almost anything without error. Now also check
that parsed URL has both a scheme and host, rejecting malformed
origins like bare paths or fragments.

Closes #336

Made-with: Cursor

* fix: remove duplicate rate limiting from handleDefault (#332)

handleDefault is already registered as the bot's default handler
which does not go through rateLimited middleware. The explicit
isRateLimited check double-consumed tokens. Removed to avoid
unfairly penalizing users during wizard flows.

Closes #332

Made-with: Cursor

* fix: sort GroupSearches output for deterministic iteration (#343)

Map iteration is non-deterministic in Go, making logs and
debugging harder. Sort groups by source, manufacturer, model.

Closes #343

Made-with: Cursor

* refactor: deduplicate listing-to-response mapping in listings.go (#345)

listListings duplicated the mapping logic already extracted in
toListingResponses in bookmarks.go. Reuse the shared helper.

Closes #345

Made-with: Cursor

* fix: add security headers to SPA handler (#339)

X-Frame-Options, X-Content-Type-Options, and Referrer-Policy
were only set on the health endpoint. Now the SPA handler also
sets them to prevent clickjacking and MIME-sniffing attacks.

Closes #339

Made-with: Cursor

* fix: only set Content-Type on requests with a body (#327)

GET and DELETE requests without a body should not send
Content-Type: application/json, which can confuse strict
servers and proxies.

Closes #327

Made-with: Cursor

* fix: increase QueryClient staleTime from 30s to 5min (#342)

30s staleTime causes excessive re-fetching on every navigation.
5 minutes is appropriate for car listing data that doesn't
change frequently.

Closes #342

Made-with: Cursor

* fix: add 404 catch-all route to prevent blank pages (#325)

Unknown URLs now show a Hebrew 404 page with a link back to
the home page instead of a blank white screen.

Closes #325

Made-with: Cursor

* fix: add ErrorBoundary to catch unhandled React exceptions (#340)

Without an ErrorBoundary, any unhandled exception crashes the
entire UI with a white screen. Now shows a Hebrew error message
with a refresh button.

Closes #340

Made-with: Cursor

* feat: add Hebrew subtitle to sidebar brand and improve mobile nav (#353, #326)

- Add "מעקב רכבים" subtitle under "CarWatch" in the sidebar
- Hide "חיפוש חדש" and "ניהול" from mobile bottom nav to reduce
  clutter (4 items instead of 6)

Closes #353, Closes #326

Made-with: Cursor

* fix: remove misleading 500MB cap and add missing table labels (#352, #354)

- Replace StorageBar with StorageIndicator that shows current
  size without implying a false 500MB limit
- Add Hebrew labels for all DB tables: seen_listings,
  catalog_cache, pending_notifications, hidden_listings,
  pending_digest, saved_listings

Closes #352, Closes #354

Made-with: Cursor

* refactor: remove dead premium grant/revoke handlers (#344)

handleGrantPremium and handleRevokePremium were defined but
never registered as bot commands. isPremium always returned
true. Simplified to a flat maxSearchesForUser and removed
the unreachable handlers. Updated tests to use defaultMaxSearches.

Closes #344

Made-with: Cursor

* fix: address CodeRabbit review comments on PR #355

- Reject CORS origins with paths, query strings, fragments, or
  userinfo — only bare scheme://host[:port] is valid
- maxSearchesForUser now respects b.maxSearches from config
  instead of always returning the hardcoded default

Made-with: Cursor

* feat: add make logs target for live log streaming

Streams docker logs from the VM with -f (follow). Supports
filtered mode via LOGS_FILTER for searching specific terms.

Made-with: Cursor

* fix: update test helpers to use defaultMaxSearches constant

Test bots hardcoded maxSearches: 3 which no longer matches
the New() default of defaultMaxSearches (10).

Made-with: Cursor

* fix: make logs filter runs grep locally to prevent shell injection

Moves grep to local side so LOGS_FILTER is never interpolated
into the remote SSH command. Also restores -f streaming flag
in filtered mode.

Made-with: Cursor

---------

+++ internal/notifier/formatter.go
@@ -56,7 +56,7 @@ func FormatListing(l model.Listing, lang locale.Lang) string {
 	}
 
 	if l.EngineVolume > 0 {
-		b.WriteString(locale.Tf(lang, "fmt_engine", l.EngineVolume/1000))
+		b.WriteString(locale.Tf(lang, "fmt_engine", float64(l.EngineVolume)/1000.0))
 		if l.GearBox != "" {
 			b.WriteString(", " + format.EscapeMarkdown(l.GearBox))
 		}
